+++ kubelet/datadog_checks/kubelet/probes.py
@@ -9,7 +9,7 @@ from datadog_checks.base.checks.kubelet_base.base import urljoin
 from datadog_checks.base.checks.openmetrics import OpenMetricsBaseCheck
 from datadog_checks.base.utils.tagging import tagger
 
-from .common import get_container_label, get_prometheus_url, replace_container_rt_prefix, tags_for_docker
+from .common import get_container_label, replace_container_rt_prefix, tags_for_docker
 
 PROBES_METRICS_PATH = '/metrics/probes'
 
@@ -28,17 +28,16 @@ class ProbesPrometheusScraperMixin(object):
             'prober_probe_total': self.prober_probe_total,
         }
 
-    def _create_probes_prometheus_instance(self, instance):
+    def _create_probes_prometheus_instance(self, instance, prom_url):
         """
         Create a copy of the instance and set default values.
         This is so the base class can create a scraper_config with the proper values.
         """
-        endpoint = get_prometheus_url("dummy_url/probes")
         probes_instance = deepcopy(instance)
         probes_instance.update(
             {
                 'namespace': self.NAMESPACE,
-                'prometheus_url': instance.get('probes_metrics_endpoint', urljoin(endpoint, PROBES_METRICS_PATH)),
+                'prometheus_url': instance.get('probes_metrics_endpoint', urljoin(prom_url, PROBES_METRICS_PATH)),
             }
         )
         return probes_instance