+++ win32_event_log/datadog_checks/win32_event_log/config_models/defaults.py
@@ -24,6 +24,10 @@ def shared_legacy_mode():
     return True
 
 
+def shared_legacy_mode_v2():
+    return False
+
+
 def shared_tag_event_id():
     return False
 
@@ -62,6 +66,10 @@ def instance_interpret_messages():
 
 def instance_legacy_mode():
     return True
+
+
+def instance_legacy_mode_v2():
+    return False
 
 
 def instance_min_collection_interval():