+++ windows_service/tests/common.py
@@ -30,7 +30,20 @@ INSTANCE_WILDCARD_DICT = {
     'disable_legacy_service_tag': True,
 }
 INSTANCE_ALL = {'services': ['ALL']}
-
+INSTANCE_PREFIX_MATCH = {
+    'services': [
+        # Include a non-name filter that should match to test that it isn't
+        # responsible for the match.
+        {'startup_type': 'automatic'},
+        # Intentionally use different letter cases to test the sorting
+        'event',
+        # The more specific filters should come last so our check must
+        # do something to ensure they are responsible for the match.
+        'EventLog',
+        'EventSystem',
+    ],
+    'disable_legacy_service_tag': True,
+}
 INSTANCE_TRIGGER_START = {
     'services': [
         {'name': 'eventlog', 'startup_type': 'automatic', 'trigger_start': False},