+++ ambari/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 <!-- towncrier release notes start -->
 
+## 4.0.0 / 2023-11-10
+
+***Changed***:
+
+* Bump the minimum base check version to 34.1.2 ([#16139](https://github.com/DataDog/integrations-core/pull/16139))
+
 ## 3.3.1 / 2023-08-18 / Agent 7.48.0
 
 ***Fixed***: