+++ ddev/CHANGELOG.md
@@ -2,6 +2,23 @@
 
 <!-- towncrier release notes start -->
 
+## 6.0.2 / 2023-11-10
+
+***Fixed***:
+
+* Add `oracledb` to the ignored dependencies when we bump them before a release ([#16155](https://github.com/DataDog/integrations-core/pull/16155))
+* Allow bumping the version of `dnspython` ([#16156](https://github.com/DataDog/integrations-core/pull/16156))
+* Add a retry mechanism when pulling the agent docker image ([#16157](https://github.com/DataDog/integrations-core/pull/16157))
+
+## 6.0.1 / 2023-11-02
+
+***Fixed***:
+
+* Fix `ddev env start` to allow the use of stable releases ([#16077](https://github.com/DataDog/integrations-core/pull/16077))
+* Stop automatically upgrading `lxml` when bumping the dependencies ([#16112](https://github.com/DataDog/integrations-core/pull/16112))
+* Properly show extra output for failed Docker Agent E2E ([#16123](https://github.com/DataDog/integrations-core/pull/16123))
+* Do not validate the codecov file in marketplace when running the `validate ci` command ([#16144](https://github.com/DataDog/integrations-core/pull/16144))
+
 ## 6.0.0 / 2023-10-26
 
 ***Changed***: