+++ win32_event_log/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 <!-- towncrier release notes start -->
 
+## 3.2.0 / 2023-11-10
+
+***Deprecated***:
+
+* Deprecate the Python implementation of the check. Setting `legacy_mode: false` now runs a check defined in [datadog-agent](https://github.com/DataDog/datadog-agent). See [datadog-agent#20701](https://github.com/DataDog/datadog-agent/pull/20701) for more information on the new check implementation. Set `legacy_mode_v2: true` to revert to the Python implementation of the check. The Python implementation may be removed in a future version of the agent. ([#16108](https://github.com/DataDog/integrations-core/pull/16108))
+
 ## 3.1.1 / 2023-10-26
 
 ***Fixed***:
@@ -9,7 +15,7 @@
 * Fix the `timeout` config option definition ([#16017](https://github.com/DataDog/integrations-core/pull/16017))
 * Set back the `timeout` option to seconds ([#16024](https://github.com/DataDog/integrations-core/pull/16024))
 
-## 3.1.0 / 2023-09-29
+## 3.1.0 / 2023-09-29 / Agent 7.49.0
 
 ***Added***: